more renaming and work on kelp
