+++ audioreactive/__init__.py
@@ -1,2 +1,5 @@
-from .lib import *
+from .util import *
+from .signal import *
+from .latent import *
+from .bend import *
 from .examples import *
